Moved transfer loop to separate module
